Added --dry-run.
Dry-run mode fetches subject.txt and calculates how many threads and posts would need to be fetched to bring the database up to date, but doesn't actually fetch them.
Incidentally, the (approximate) number of posts that will be fetched is now displayed prior to scraping in regular mode as well.