Let's pay closer attention to HTTP status codes
This change is specifically to give better feedback to banned users. As it was, /prog/scrape quietly fetched subject.txt, didn't notice it didn't actually get it (it got a blank body), and just reported that there were no new posts. Now, it will display a ``! Error: 302 Banned'' message (on world4ch, at least) and exit.