Now handles chunked transfer-encoding properly.
When using Transfer-Encoding chunked, httplib's HTTPResponse can only be expected to contain headers, apparently, as HTTPConnection.getresponse() returns without waiting for the entire body. If you close the connection before doing a read(), which does wait until the last bit of the reply is received before returning, you will probably have only part or none of the body.
World4ch recently switched to Transfer-Encoding: chunked. This is what was going wrong, and why /prog/scrape thought subject.txt was empty and the database never needed updating.

I can see why the httplib developers could have considered this to be a reasonable way to do things, but given the lack of documentation, I still consider this behaviour to be a bug in httplib.