Just skip failed threads instead of aborting.
This is a temporary work-around and a bit of a dirty hack. Though just skipping the thread probably makes more sense than just terminating the scraper thread anyway; errors during scraping haven't terminated the whole application since the move to multi-threading.

I don't know how he managed it, but the recent world4ch configuration changes have really fucked /prog/scrape in the ass. On the other hand, there's no excuse for httplib to be this intolerant to shitty web servers either. Maybe switching back to urllib2 is the easiest fix.