Switched back to urllib2.
Too bad, but httplib is too fragile in the face of 4chan sysadmins. This change fixes all of the occasional errors loading pages, at the expense of keep-alive connections.
It's also noticeably faster than it was before, but whether this is because of MVB's configuration changes or because of httplib, I don't know.